PerlinNoise/Turbulence : Remove
These are inferior to the implementations provided by OpenShadingLanguage in `liboslnoise`, and are not used anywhere in Cortex or Gaffer. They also relied on `std::random_shuffle()`, which is not consistent between platforms and was deprecated in C++14 and removed in C++17. We can't use the just-hardcode-the-table approach to backwards compatibility we used with PointDistribution, because PerlinNoise generates a different table on the fly each time, using the seed provided.